// app/sdms-v4/src/apis/info/model.ts
// ============================================================================
// Copyright 2017-2024, Schlumberger
//
// Licensed under the Apache License, Version 2.0 (the "License");
// You may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// Distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// Limitations under the License.
// ============================================================================

export interface Info {
    group_id: string;
    artifact_id: string;
    version?: string;
    build_time?: string;
    branch?: string;
    commit_id?: string;
    commit_message?: string;
    connected_outer_services?: string[];
}